the room services has been completed but not tested  and yet to be mapped correctly

+++ src/main/java/com/sevenStar/hotel/models/repositories/AdminRepository.java
@@ -3,5 +3,5 @@ package com.sevenStar.hotel.models.repositories;
 import com.sevenStar.hotel.models.entities.AdminUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminRepository extends JpaRepository<AdminUser, Integer> {
+public interface AdminRepository extends JpaRepository<AdminUser, Long> {
 }
